test(backend): move data sharing tests to module structure
